docs(refactor): split the first roadmap item into two
