+++ mainline-next/docs/1a/api.rs
@@ -1,5 +1,6 @@
+// Roadmap milestone 1a: minimal safe IPv4 client and low-level API.
+
 use std::{
-    future::Future,
     net::{Ipv4Addr, SocketAddrV4},
     num::NonZeroUsize,
     time::{Duration, Instant},
@@ -7,7 +8,7 @@ use std::{
 
 use futures_lite::Stream;
 
-// Configuration, bootstrap, health and lifecycle shared by both API levels.
+// Configuration, bootstrap, health and lifecycle used by both API levels.
 // Dht is cloneable. Dht handles and active streams keep the library-owned Mio
 // reactor alive; dropping the last holder stops and joins it.
 
@@ -117,16 +118,23 @@ pub struct NetworkActivity {
     pub timed_out: usize,
 }
 
-pub enum BootstrapStatus {
-    Resolving(BootstrapResolutionProgress),
-    InProgress(BootstrapProgress),
-    Complete(BootstrapOutcome),
+pub struct BootstrapStatus {
+    // DNS resolution and DHT traversal may progress at the same time.
+    pub resolution: BootstrapResolution,
+    pub traversal: BootstrapTraversal,
 }
 
-pub struct BootstrapResolutionProgress {
+pub struct BootstrapResolution {
     pub pending_names: usize,
+    pub resolved_names: usize,
     pub resolved_addresses: usize,
     pub failed_names: usize,
+}
+
+pub enum BootstrapTraversal {
+    NotStarted,
+    InProgress(BootstrapProgress),
+    Complete(BootstrapTraversalOutcome),
 }
 
 pub struct BootstrapProgress {
@@ -136,8 +144,12 @@ pub struct BootstrapProgress {
 }
 
 pub struct BootstrapOutcome {
+    pub resolution: BootstrapResolution,
+    pub traversal: BootstrapTraversalOutcome,
+}
+
+pub struct BootstrapTraversalOutcome {
     pub state: DhtHealthState,
-    pub dns_failures: usize,
     pub queried: usize,
     pub responded: usize,
     pub timed_out: usize,
@@ -354,6 +366,11 @@ pub struct MutableGetReport {
     pub requests_sent: usize,
     pub responses_received: usize,
     pub unique_responders: usize,
+    // Cumulative validated, unique response categories across the lookup.
+    // These are diagnostics, not current closest-set coverage.
+    pub item_responses: usize,
+    pub no_value_responses: usize,
+    pub no_more_recent_responses: usize,
     pub rejected_responses: usize,
     pub protocol_errors: usize,
     pub final_closest_nodes: usize,
@@ -365,195 +382,4 @@ pub enum QueryError {
     AdmissionTimeout,
     OverallDeadline,
     Shutdown,
-}
-
-// High-level API: adapters over only the public low-level streams above. This
-// logic requires no access to the reactor, routing table, or RPC internals.
-
-impl Dht {
-    pub async fn get_mutable(
-        &self,
-        public_key: &[u8; 32],
-        salt: Option<&[u8]>,
-        more_recent_than: Option<i64>,
-    ) -> Result<MutableEstimateStream, QueryError> {
-        let responses = self
-            .get_mutable_responses(public_key, salt, more_recent_than)
-            .await?;
-        Ok(responses.into_estimates())
-    }
-
-    pub async fn get_mutable_with_policy(
-        &self,
-        public_key: &[u8; 32],
-        salt: Option<&[u8]>,
-        more_recent_than: Option<i64>,
-        policy: MutableGetPolicy,
-    ) -> Result<MutableEstimateStream, QueryError> {
-        let responses = self
-            .get_mutable_responses(public_key, salt, more_recent_than)
-            .await?;
-        Ok(responses.into_estimates_with_policy(policy))
-    }
-
-    pub async fn put_mutable(
-        &self,
-        item: MutableItem,
-    ) -> Result<MutablePutConclusion, QueryError> {
-        let events = self.put_mutable_events(item).await?;
-        Ok(events.into_conclusion().await)
-    }
-}
-
-impl MutableLookupStream {
-    // Uses MutableGetPolicy::Adaptive.
-    pub fn into_estimates(self) -> MutableEstimateStream;
-    pub fn into_estimates_with_policy(self, policy: MutableGetPolicy)
-        -> MutableEstimateStream;
-}
-
-impl MutablePutStream {
-    pub fn into_conclusion(self) -> MutablePutConclusionFuture;
-}
-
-pub struct MutablePutConclusionFuture {
-    // Private fields contain only a MutablePutStream and adapter state.
-}
-
-impl Future for MutablePutConclusionFuture {
-    type Output = MutablePutConclusion;
-}
-
-pub enum MutablePutConclusion {
-    Published(MutablePutPublished),
-    Conflict(MutablePutConflict),
-    Inconclusive(MutablePutInconclusive),
-}
-
-// Private fields keep conclusion-specific invariants intact. In particular, a
-// Published conclusion always has at least one acknowledgement.
-pub struct MutablePutPublished {
-    // Private fields.
-}
-
-impl MutablePutPublished {
-    pub fn acknowledgements(&self) -> NonZeroUsize;
-    pub fn evidence(&self) -> &MutablePutEvidence;
-}
-
-pub struct MutablePutConflict {
-    // Private fields.
-}
-
-impl MutablePutConflict {
-    pub fn newer(&self) -> &MutableItem;
-    pub fn evidence(&self) -> &MutablePutEvidence;
-}
-
-pub struct MutablePutInconclusive {
-    // Private fields.
-}
-
-impl MutablePutInconclusive {
-    pub fn reason(&self) -> &MutablePutInconclusiveReason;
-    pub fn evidence(&self) -> &MutablePutEvidence;
-}
-
-pub enum MutablePutInconclusiveReason {
-    NoAcknowledgement,
-    QueryTimeout,
-    OverallDeadline,
-    Unreachable,
-    Shutdown,
-    ProtocolError,
-}
-
-pub struct MutablePutEvidence {
-    // Private fields derived solely from MutablePutEvent values.
-}
-
-impl MutablePutEvidence {
-    pub fn report(&self) -> &MutablePutReport;
-    pub fn acknowledgements(&self) -> usize;
-    pub fn attempted_targets(&self) -> usize;
-    pub fn target_set_size(&self) -> usize;
-}
-
-pub struct MutableEstimateStream {
-    // Private fields contain only a MutableLookupStream and adapter state.
-    // It therefore has the same stream-driven execution and cancellation.
-}
-
-impl Stream for MutableEstimateStream {
-    type Item = MutableEstimateUpdate;
-}
-
-pub enum MutableGetPolicy {
-    // Return after traversal, sufficient relative coverage, and adaptive
-    // settling. Slow stragglers need not hold the query open.
-    Adaptive,
-
-    // Wait for every relevant request to receive a response or time out.
-    Strict,
-}
-
-pub struct MutableEstimateUpdate {
-    pub evidence: MutableGetEvidence,
-    pub status: MutableGetStatus,
-}
-
-pub enum MutableGetStatus {
-    Searching {
-        estimate: Option<MutableEstimate>,
-    },
-    Converged {
-        estimate: MutableEstimate,
-        report: MutableGetReport,
-    },
-    NotFound {
-        report: MutableGetReport,
-    },
-    Inconclusive {
-        estimate: Option<MutableEstimate>,
-        reason: MutableGetInconclusiveReason,
-        report: MutableGetReport,
-    },
-}
-
-pub struct MutableEstimate {
-    // Private fields preserve consistency with the accompanying evidence.
-}
-
-impl MutableEstimate {
-    pub fn item(&self) -> &MutableItem;
-    pub fn supporting_nodes(&self) -> usize;
-}
-
-pub struct MutableGetEvidence {
-    // Private fields derived from one MutableLookupProgress snapshot.
-}
-
-impl MutableGetEvidence {
-    // Coverage values refer only to unique valid responses from the current
-    // closest set. Together they fully explain the adapter's decision.
-    pub fn covered_nodes(&self) -> usize;
-    pub fn coverage_basis(&self) -> usize;
-    pub fn required_coverage(&self) -> usize;
-    pub fn allowed_outstanding(&self) -> usize;
-
-    pub fn closest_nodes(&self) -> usize;
-    pub fn pending(&self) -> usize;
-    pub fn failed(&self) -> usize;
-    pub fn unchanged_for(&self) -> Duration;
-    pub fn settling_period(&self) -> Duration;
-    pub fn traversal_converged(&self) -> bool;
-}
-
-pub enum MutableGetInconclusiveReason {
-    QueryTimeout,
-    OverallDeadline,
-    Unreachable,
-    Shutdown,
-    ProtocolError,
-    InsufficientCoverage,
 }
