+++ mainline-next/docs/README.md
@@ -2,10 +2,22 @@
 
 This proposal exposes low-level mutable GET and PUT event streams and implements
 the high-level estimate and conclusion APIs only as adapters over those streams.
+The documents are organized around roadmap milestones 1a and 1b.
 
 - [Problem statement](problem-statement.md)
 - [Design principles](design-principles.md)
-- [Brief requirements](requirements-brief.md)
-- [Requirements](requirements.md)
-- [API proposal](api.rs)
-- [Prototype plan](prototype-plan.md)
+- [Roadmap](roadmap.md)
+
+## 1a. Minimal Safe IPv4 Client and Low-Level API
+
+- [Brief requirements](1a/requirements-brief.md)
+- [Requirements](1a/requirements.md)
+- [API proposal](1a/api.rs)
+- [Prototype plan](1a/prototype-plan.md)
+
+## 1b. High-Level IPv4 Item API
+
+- [Brief requirements](1b/requirements-brief.md)
+- [Requirements](1b/requirements.md)
+- [API proposal](1b/api.rs)
+- [Prototype plan](1b/prototype-plan.md)
